refactor(platform-api): extract orgs service into domain
